+++ _examples/file-server/single-page-application/basic/main.go
@@ -28,6 +28,8 @@ func newApp() *iris.Application {
 	// })
 
 	assetHandler := app.StaticHandler("./public", false, false)
+	// as an alternative of SPA you can take a look at the /routing/dynamic-path/root-wildcard
+	// example too
 	app.SPA(assetHandler)
 
 	return app